requirements.txt
pyqt5
setuptools
tzdata
wheel
pillow
numpy
